% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{diagnose_report}
\alias{diagnose_report}
\title{Reporting the information of data diagnosis with html}
\usage{
diagnose_report(
  .data,
  output_file = NULL,
  output_dir = tempdir(),
  browse = TRUE,
  title = "Data Diagnosis",
  subtitle = deparse(substitute(.data)),
  author = "",
  title_color = "gray",
  thres_uniq_cat = 0.5,
  thres_uniq_num = 5,
  logo_img = NULL,
  create_date = format(Sys.Date(), "\%d \%B \%Y"),
  theme = c("orange", "blue")[1],
  sample = 1,
  ...
)
}
\arguments{
\item{.data}{a data.frame or a \code{\link{tbl_df}}.}

\item{output_file}{name of generated file. default is NULL.}

\item{output_dir}{name of directory to generate report file. default is tempdir().}

\item{browse}{logical. choose whether to output the report results to the browser.}

\item{title}{character. title of report. default is "Data Diagnosis Report".}

\item{subtitle}{character. subtitle of report. default is name of data.}

\item{author}{character. author of report. default is "".}

\item{title_color}{character. color of title. default is "white".}

\item{thres_uniq_cat}{numeric. threshold to use for "Unique Values - 
Categorical Variables". default is 0.5.}

\item{thres_uniq_num}{numeric. threshold to use for "Unique Values - 
Numerical Variables". default is 5.}

\item{logo_img}{character. name of logo image on top right.}

\item{theme}{character. name of theme for report. support "orange" and "blue". 
default is "orange".}

\item{sample}{numeric. ratio of the sample. It takes a real number 
between (0, 1) as an argument value. The default value is 1, and all data 
is diagnosed.}

\item{...}{arguments to be passed to methods.}
}
\description{
The diagnose_report() report the information for diagnosing
the quality of the data.
}
\details{
Generate generalized data diagnostic reports automatically.
You can choose to output to pdf and html files.
This is useful for diagnosing a data frame with a large number of variables
than data with a small number of variables.

The title and subtitle colors should be designated by the color name used in 
CSS of html, not the color defined in R.
}
\section{Reported information}{

Reported from the data diagnosis is as follows.

\itemize{
  \item Data Diagnosis
  \itemize{
    \item Overview 
    \itemize{
      \item Missing/Unique Values
    }
    \item Missing Values
    \itemize{
      \item List of Missing Values
      \item Visualization
    }
    \item Unique Values
    \itemize{
      \item Categorical Variables
      \item Numerical Variables
    }
  }
  \item Categorical Variable Diagnosis
  \itemize{
    \item Top Ranks
  }   
  \item Numerical Variable Diagnosis
  \itemize{
    \item Distribution
    \itemize{
      \item Zero Values
      \item Minus Values
    }
    \item Outliers
    \itemize{
      \item List of Outliers
      \item Individual Outliers
    }
  }
}
}

\examples{
\dontrun{
# create pdf file. file name is Diagnosis_Paged_Report.pdf
diagnose_report(heartfailure)

# file name is Diagn.html. and change logo image
# logo <- file.path(system.file(package = "dlookrExtra"), "report", "R_logo_html.svg")
# diagnose_report(heartfailure, logo_img = logo, title_color = "black",
    output_file = "Diagn.html")

# file name is ./Diagn_heartfailure.html, "blue" theme and not browse
# diagnose_report(heartfailure, output_dir = ".", author = "Choonghyun Ryu",
#    output_file = "Diagn_heartfailure.html", theme = "blue", browse = FALSE)
}

}
