% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlookrExtra_templ_paged.R
\name{dlookrExtra_templ_paged}
\alias{dlookrExtra_templ_paged}
\title{Generate paged HTML document}
\usage{
dlookrExtra_templ_paged(...)
}
\arguments{
\item{...}{arguments to be passed to \code{pagedown::\link[pagedown]{html_paged}}.}
}
\value{
document of markdown format.
}
\description{
Generate paged HTML document
}
\references{
\url{https://pagedown.rbind.io}
}
