% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda.R
\name{eda_category}
\alias{eda_category}
\title{EDA report of categorical variables}
\usage{
eda_category(
  x,
  target_variable,
  positive = "1",
  output_file = NULL,
  output_dir = getwd(),
  sample_percent = 100,
  parallel = FALSE,
  cores = parallel::detectCores() - 2,
  future_globals_maxsize = 500 * 1024^2,
  base_family = NULL,
  verbose = TRUE
)
}
\arguments{
\item{x}{data.frame, tbl_df. Dataset to perform EDA on.}

\item{target_variable}{character. The name of the target variable.}

\item{positive}{character. The name of the positive class in the target variable}

\item{output_file}{character. The name of the excel file to be created.}

\item{output_dir}{character. The directory where the excel file to be created will be stored.}

\item{sample_percent}{numeric. Percentage of sample when analyzed by sample.}

\item{parallel}{logical. Whether to perform parallel processing, defaults to FALSE.}

\item{cores}{integer. The number of cores to use for parallel processing.}

\item{future_globals_maxsize}{numeric. Maximum amount of global memory to 
be used by individual cores in parallel processing.}

\item{base_family}{character. Font family name to use for visualization.}

\item{verbose}{logical. Whether to output job progress information, 
the default value is TRUE.}
}
\description{
Generating an excel EDA report of categorical variables
}
\details{
EDA information is output on one worksheet for all categorical 
variables except for the overall information of categorical variables.
Excel only supports 256 worksheets, so anything with more than 255 categorical 
variables produces only 255 pieces of information.
}
\examples{
\dontrun{
testdata <- dlookr::heartfailure
target_variable <- "death_event"

# single core processing
eda_category(testdata, target_variable, positive = "Yes")

# Generate reports with 30\% sampled data
eda_category(testdata, target_variable, positive = "Yes", sample_percent = 30)

# parallel processing
eda_category(testdata, target_variable, positive = "Yes", parallel = TRUE,
             cores = 8)
}

}
\seealso{
\code{\link{eda_numeric}}
}
