% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda.R
\name{eda_numeric}
\alias{eda_numeric}
\title{EDA report of numerical variables}
\usage{
eda_numeric(
  x,
  target_variable,
  positive = "1",
  output_file = NULL,
  output_dir = getwd(),
  sample_percent = 100,
  trim_quantile = NULL,
  parallel = FALSE,
  cores = parallel::detectCores() - 2,
  future_globals_maxsize = 500 * 1024^2,
  base_family = NULL,
  verbose = TRUE
)
}
\arguments{
\item{x}{data.frame, tbl_df. EDA를 수행할 데이터셋.}

\item{target_variable}{character. The name of the target variable.}

\item{positive}{character. The name of the positive class in the target variable}

\item{output_file}{character. The name of the excel file to be created.}

\item{output_dir}{character. The directory where the excel file to be created will be stored.}

\item{sample_percent}{numeric. Percentage of sample when analyzed by sample.}

\item{trim_quantile}{numeric. The upper and lower quantiles (%) to trim the 
data to remove outliers.
The default is NULL to generate a report of individual continuous variables 
for all data, i.e. [min, max].
In the case of a numeric vector of length 2, the lower and upper bounds are expressed as percentiles.
For example, in the case of c(0, 99), data in the range of [max, 99th percentile] is targeted, 
and in the case of c(5, 95), data in the range of [5th percentile, 95th percentile] 
are individually selected. Generates reports of continuous variables.
If this value is NA, reports are made for individual continuous variables for 
values in the range (Q1 - 1.5 * IQR, Q3 + 1.5 * IQR). 
In other words, a report of individual continuous variables is generated 
with the data with outliers removed.}

\item{parallel}{logical. Whether to perform parallel processing, defaults to FALSE.}

\item{cores}{integer. The number of cores to use for parallel processing.}

\item{future_globals_maxsize}{numeric. Maximum amount of global memory to 
be used by individual cores in parallel processing.}

\item{base_family}{character. Font family name to use for visualization.}

\item{verbose}{logical. Whether to output job progress information, 
the default value is TRUE.}
}
\description{
EDA performance results for continuous variables are created as an excel file.
}
\details{
EDA information is output on one worksheet for all numerical 
variables except for the overall information of numerical variables.
Excel only supports 256 worksheets, so anything with more than 255 numerical 
variables produces only 255 pieces of information.
}
\examples{
\dontrun{
testdata <- dlookr::heartfailure
target_variable <- "death_event"

# single core processing
eda_numeric(testdata, target_variable, positive = "Yes")

# Generate reports with 30\% sampled data
eda_numeric(testdata, target_variable, positive = "Yes", sample_percent = 30)

# Create a report of individual variables with data with outliers removed
eda_numeric(testdata, target_variable, positive = "Yes", trim_quantile = NA)

# Create a report of individual variables with data with 5\% of the upper data limit removed
eda_numeric(testdata, target_variable, positive = "Yes", trim_quantile = c(0, 95))

# parallel processing
eda_numeric(testdata, target_variable, positive = "Yes", parallel = TRUE,
            cores = 8)
}

}
\seealso{
\code{\link{eda_category}}
}
