% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{eda_report}
\alias{eda_report}
\title{Reporting the information of EDA with html}
\usage{
eda_report(
  .data,
  target = NULL,
  output_file = NULL,
  output_dir = tempdir(),
  browse = TRUE,
  title = "EDA",
  subtitle = deparse(substitute(.data)),
  author = "dlookr",
  title_color = "gray",
  logo_img = NULL,
  create_date = Sys.time(),
  theme = c("orange", "blue")[1],
  sample_percent = 100,
  ...
)
}
\arguments{
\item{.data}{a data.frame or a \code{\link{tbl_df}}.}

\item{output_file}{name of generated file. default is NULL.}

\item{output_dir}{name of directory to generate report file. default is tempdir().}

\item{browse}{logical. choose whether to output the report results to the browser.}

\item{title}{character. title of report. default is "Data Diagnosis Report".}

\item{subtitle}{character. subtitle of report. default is name of data.}

\item{author}{character. author of report. default is "dlookr".}

\item{title_color}{character. color of title. default is "white".}

\item{logo_img}{character. name of logo image on top right.}

\item{theme}{character. name of theme for report. support "orange" and "blue". 
default is "orange".}

\item{sample_percent}{numeric. Sample percent of data for performing EDA. 
It has a value between (0, 100]. 100 means all data, and 5 means 5% of sample data.
This is useful for data with a large number of observations.}

\item{...}{arguments to be passed to methods.}

\item{thres_uniq_cat}{numeric. threshold to use for "Unique Values - 
Categorical Variables". default is 0.5.}

\item{thres_uniq_num}{numeric. threshold to use for "Unique Values - 
Numerical Variables". default is 0.1.}
}
\description{
The eda_report() report the information for diagnosing
the quality of the data.
}
\details{
Generate generalized data diagnostic reports automatically.
You can choose to output to pdf and html files.
This is useful for diagnosing a data frame with a large number of variables
than data with a small number of variables.

The title and subtitle colors should be designated by the color name used in 
CSS of html, not the color defined in R.
}
\section{Reported information}{

Reported from the data diagnosis is as follows.

\itemize{
  \item Data Diagnosis
  \itemize{
    \item Overview 
    \item Missing Values
    \itemize{
      \item List of Missing Values
      \item Visualization
    }
    \item Unique Values
    \itemize{
      \item Categorical Variables
      \item Numerical Variables
    }
  }
  \item Categorical Variable Diagnosis
  \itemize{
    \item Top Ranks
  }   
  \item Numerical Variable Diagnosis
  \itemize{
    \item Distribution
    \itemize{
      \item Zero Values
      \item Minus Values
    }
    \item Outliers
    \itemize{
      \item List of Outliers
      \item Individual Outliers
    }
  }
}
}

\examples{
\dontrun{
# create pdf file. file name is Diagnosis_Paged_Report.pdf
eda_report(heartfailure)

# file name is EDA.html. and change logo image
# logo <- file.path(system.file(package = "dlookrExtra"), "report", "R_logo_html.svg")
# eda_report(heartfailure, logo_img = logo, title_color = "black",
    output_file = "EDA.html")

# file name is ./EDA_heartfailure.html, "blue" theme and not browse
# eda_report(heartfailure, output_dir = ".", author = "Choonghyun Ryu",
#    output_file = "EDA_heartfailure.html", theme = "blue", browse = FALSE)
}

}
