% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paged.R
\name{break_page}
\alias{break_page}
\alias{break_page_asis}
\alias{break_line_asis}
\title{Breaking pages in HTML}
\usage{
break_page()

break_page_asis()

break_line_asis(n = 1)
}
\value{
HTML code
}
\description{
Breaking pages in HTML
}
\details{
break_page() insert the HTML page break code in the R code chunk. and 
break_page_asis() inserts the HTML page break code in the asis statement of 
the R code chunk.
}
